input-ex: use same renderer, remove styled…
