+++ src/components/content/exercises/exercise.tsx
@@ -57,9 +57,7 @@ export function Exercise({ node, renderNested, path }: ExerciseProps) {
 
   const auth = useAuthentication()
   const [loaded, setLoaded] = useState(false)
-  useEffect(() => {
-    setLoaded(true)
-  }, [])
+  useEffect(() => setLoaded(true), [])
 
   const { asPath } = useRouter()
