+++ src/components/content/exercises/input-exercise.tsx
@@ -1,14 +1,10 @@
-import type A from 'algebra.js'
-import clsx from 'clsx'
 import { useRouter } from 'next/router'
-import { useState, useEffect } from 'react'
 
-import { Feedback } from './feedback'
 import { useInstanceData } from '@/contexts/instance-context'
 import { EdtrPluginInputExercise } from '@/frontend-node-types'
 import { exerciseSubmission } from '@/helper/exercise-submission'
-import { tw } from '@/helper/tw'
 import { RenderNestedFunction } from '@/schema/article-renderer'
+import { InputExerciseRenderer } from '@/serlo-editor/plugins/input-exercise/renderer'
 
 export interface InputExerciseProps {
   data: EdtrPluginInputExercise['state']
@@ -20,138 +16,49 @@ export interface InputExerciseProps {
   }
 }
 
-interface FeedbackData {
-  correct: boolean
-  message: JSX.Element
-}
-
 export function InputExercise({
   data,
   renderNested,
   isRevisionView,
   context,
 }: InputExerciseProps) {
-  const [feedback, setFeedback] = useState<FeedbackData | null>(null)
-  const [value, setValue] = useState('')
-  const [A, setA] = useState<typeof import('algebra.js') | null>(null)
   const exStrings = useInstanceData().strings.content.exercises
-
   const { asPath } = useRouter()
 
-  useEffect(() => {
-    void import('algebra.js').then((value) => setA(value))
-  }, [])
+  return (
+    <>
+      <InputExerciseRenderer
+        type={data.type}
+        unit={data.unit}
+        answers={renderAnswers()}
+        onEvaluate={onEvaluate}
+      />
+      {isRevisionView && renderRevisionExtra()}
+    </>
+  )
 
-  function evaluate() {
-    const feedbackData = checkAnswer()
+  function onEvaluate(correct: boolean) {
     exerciseSubmission({
       path: asPath,
       entityId: context.entityId,
       revisionId: context.revisionId,
-      result: feedbackData.correct ? 'correct' : 'wrong',
+      result: correct ? 'correct' : 'wrong',
       type: 'input',
     })
-    setFeedback(feedbackData)
   }
 
-  return (
-    <div className="mx-side mb-7">
-      <input
-        className={tw`
-            serlo-input-font-reset mb-5
-            rounded-3xl border-3 border-brand-400 py-2
-            px-3 font-bold text-brand placeholder-brand focus:border-brand focus:bg-white
-            focus:text-brand focus:placeholder-opacity-0 focus:opacity-100 focus:outline-none
-            active:border-brand print:hidden
-          `}
-        type="text"
-        value={value}
-        onChange={(e) => setValue(e.target.value)}
-        onKeyDown={(e) => {
-          if (e.key === 'Enter') evaluate()
-        }}
-        placeholder={exStrings.yourAnswer}
-      />{' '}
-      {data.unit}
-      <br />
-      {feedback && (
-        <Feedback correct={feedback.correct}>{feedback.message}</Feedback>
-      )}
-      {A && (
-        <>
-          <a
-            className={clsx(
-              'serlo-button-blue mt-4',
-              value === '' && 'pointer-events-none opacity-0'
-            )}
-            onClick={evaluate}
-          >
-            {exStrings.check}
-          </a>
-          {isRevisionView && renderRevisionExtra()}
-        </>
-      )}
-    </div>
-  )
+  function renderAnswers() {
+    return data.answers.map((answer) => {
+      const isEmpty =
+        !answer.feedback[0].children ||
+        (answer.feedback[0].children?.length === 1 &&
+          answer.feedback[0].children[0].children?.length === 0)
 
-  function checkAnswer(): FeedbackData {
-    const answers = data.answers
-    const filteredAnswers = answers.filter((answer) => {
-      try {
-        const solution = normalize(answer.value)
-        const submission = normalize(value)
-
-        if (
-          data.type === 'input-expression-equal-match-challenge' &&
-          solution
-        ) {
-          return (
-            (solution as A.Expression)
-              .subtract(submission as A.Expression)
-              .toString() === '0'
-          )
-        }
-        return solution === submission
-      } catch (e) {
-        return false
+      return {
+        ...answer,
+        feedback: isEmpty ? null : <>{renderNested(answer.feedback)}</>,
       }
     })
-    const hasCustomFeedback = filteredAnswers[0]?.feedback.length > 0
-    const customFeedbackNode =
-      hasCustomFeedback &&
-      (renderNested(filteredAnswers[0]?.feedback, 'feedback') as JSX.Element)
-
-    if (filteredAnswers.length < 1 || !filteredAnswers[0]?.isCorrect) {
-      return {
-        correct: false,
-        message: customFeedbackNode || <>{exStrings.wrong}</>,
-      }
-    } else {
-      return {
-        correct: true,
-        message: customFeedbackNode || <>{exStrings.correct}</>,
-      }
-    }
-  }
-
-  function normalize(value: string) {
-    const _value = collapseWhitespace(value)
-    switch (data.type) {
-      case 'input-number-exact-match-challenge':
-        return normalizeNumber(_value).replace(/\s/g, '')
-      case 'input-expression-equal-match-challenge':
-        return A ? A.parse(normalizeNumber(_value)) : undefined
-      case 'input-string-normalized-match-challenge':
-        return _value.toUpperCase()
-    }
-  }
-
-  function collapseWhitespace(val: string): string {
-    return val.replace(/[\s\xa0]+/g, ' ').trim()
-  }
-
-  function normalizeNumber(val: string) {
-    return val.replace(/,/g, '.').replace(/^[+]/, '')
   }
 
   function renderRevisionExtra() {
