+++ src/serlo-editor/plugins/input-exercise/renderer.tsx
@@ -1,185 +1,131 @@
-import A from 'algebra.js'
-import { useState, createRef } from 'react'
-import styled from 'styled-components'
+import type A from 'algebra.js'
+import clsx from 'clsx'
+import { useState, useEffect } from 'react'
 
-import { InputExerciseProps, InputExerciseType } from '.'
-import { Feedback, SubmitButton } from '../../renderer-ui'
-import { store, selectIsDocumentEmpty } from '../../store'
+import { Feedback } from '@/components/content/exercises/feedback'
 import { useInstanceData } from '@/contexts/instance-context'
-import { useEditorStrings } from '@/contexts/logged-in-data-context'
-import { legacyEditorTheme } from '@/helper/colors'
+import { tw } from '@/helper/tw'
 
-enum ExerciseState {
-  Default = 1,
-  SolvedRight,
-  SolvedWrong,
+interface InputExersiseRendererProps {
+  type: string
+  unit: string
+  answers: {
+    value: string
+    isCorrect: boolean
+    feedback: JSX.Element | null
+  }[]
+  onEvaluate?: (correct: boolean) => void
 }
 
-const InputContainer = styled.div({
-  float: 'right',
-  display: 'flex',
-  flexDirection: 'row',
-})
+interface FeedbackData {
+  correct: boolean
+  message: JSX.Element
+}
 
-const InputExerciseField = styled.input({
-  border: 'none',
-  borderBottom: `3px solid ${legacyEditorTheme.primary.background}`,
-  textAlign: 'center',
-  outline: 'none',
-  marginBottom: '10px',
-})
+export function InputExerciseRenderer({
+  type,
+  unit,
+  answers,
+  onEvaluate,
+}: InputExersiseRendererProps) {
+  const [feedback, setFeedback] = useState<FeedbackData | null>(null)
+  const [value, setValue] = useState('')
+  const [A, setA] = useState<typeof import('algebra.js') | null>(null)
+  const exStrings = useInstanceData().strings.content.exercises
 
-export function InputExerciseRenderer(props: InputExerciseProps) {
-  const { state } = props
-  const [feedbackIndex, setFeedbackIndex] = useState<number>(-1)
-  const [feedbackVisible, setFeedbackVisible] = useState<boolean>()
-  const [exerciseState, setExerciseState] = useState<ExerciseState>(
-    ExerciseState.Default
-  )
-  const editorStrings = useEditorStrings()
-  const { strings } = useInstanceData()
+  useEffect(() => void import('algebra.js').then((value) => setA(value)), [])
 
-  const input = createRef<HTMLInputElement>()
-  const handleWrongAnswer = () => {
-    setTimeout(() => {
-      setExerciseState(ExerciseState.Default)
-    }, 2000)
-    setExerciseState(ExerciseState.SolvedWrong)
-  }
-
-  function checkAnswer(event: React.FormEvent) {
-    if (!input.current) {
-      return
-    }
-    event.preventDefault()
-    setFeedbackIndex(-1)
-    setFeedbackVisible(true)
-    setExerciseState(ExerciseState.Default)
-    if (input.current.value === '') {
-      setFeedbackVisible(false)
-      return
-    }
-    const { state } = props
-
-    let containedAnswer = false
-    state.answers.forEach((answer, index) => {
-      if (
-        input.current &&
-        matchesInput(
-          {
-            type: state.type.value as InputExerciseType,
-            value: answer.value.value,
-          },
-          input.current.value
-        )
-      ) {
-        setFeedbackIndex(index)
-        if (answer.isCorrect.value) {
-          setExerciseState(ExerciseState.SolvedRight)
-        } else {
-          handleWrongAnswer()
-        }
-        containedAnswer = true
-      }
-    })
-
-    if (!containedAnswer) {
-      handleWrongAnswer()
-    }
+  function evaluate() {
+    const feedbackData = checkAnswer()
+    if (onEvaluate) onEvaluate(feedbackData.correct)
+    setFeedback(feedbackData)
   }
 
   return (
-    <div>
-      <form onSubmit={checkAnswer}>
-        <InputContainer>
-          <InputExerciseField
-            onKeyDown={(k: React.KeyboardEvent<HTMLInputElement>) => {
-              const { key } = k as unknown as KeyboardEvent
-              if ((key === 'Enter' || key === 'Backspace') && props.editable) {
-                k.stopPropagation()
-              }
-            }}
-            data-type={state.type.value}
-            type="text"
-            placeholder={
-              editorStrings.templatePlugins.inputExercise.yourSolution
-            }
-            ref={input}
-          />
-          {state.unit.value}
-        </InputContainer>
-        <div className="clear-both" />
-
-        {feedbackVisible ? (
-          feedbackIndex > -1 ? (
-            <Feedback
-              boxFree
-              isTrueAnswer={state.answers[feedbackIndex].isCorrect.value}
-            >
-              {selectIsDocumentEmpty(
-                store.getState(),
-                state.answers[feedbackIndex].feedback.id
-              )
-                ? state.answers[feedbackIndex].isCorrect.value
-                  ? strings.content.exercises.correct
-                  : strings.content.exercises.wrong
-                : state.answers[feedbackIndex].feedback.render()}
-            </Feedback>
-          ) : (
-            <Feedback boxFree>{strings.content.exercises.wrong}</Feedback>
-          )
-        ) : null}
-        <div>
-          <SubmitButton exerciseState={exerciseState} />
-          <div className="clear-both" />
-        </div>
-      </form>
+    <div className="mx-side mb-7">
+      <input
+        className={tw`
+            serlo-input-font-reset mb-5
+            rounded-3xl border-3 border-brand-400 py-2
+            px-3 font-bold text-brand placeholder-brand focus:border-brand focus:bg-white
+            focus:text-brand focus:placeholder-opacity-0 focus:opacity-100 focus:outline-none
+            active:border-brand print:hidden
+          `}
+        type="text"
+        value={value}
+        onChange={(e) => setValue(e.target.value)}
+        onKeyDown={(e) => {
+          if (e.key === 'Enter') evaluate()
+        }}
+        placeholder={exStrings.yourAnswer}
+      />{' '}
+      {unit}
+      <br />
+      {feedback && (
+        <Feedback correct={feedback.correct}>{feedback.message}</Feedback>
+      )}
+      {A && (
+        <a
+          className={clsx(
+            'serlo-button-blue mt-4 !text-white',
+            value === '' && 'pointer-events-none opacity-0'
+          )}
+          onClick={evaluate}
+        >
+          {exStrings.check}
+        </a>
+      )}
     </div>
   )
-}
 
-function normalize(type: InputExerciseType, text: string) {
-  const temp = collapseWhitespace(text)
+  function checkAnswer(): FeedbackData {
+    const filteredAnswers = answers.filter((answer) => {
+      try {
+        const solution = normalize(answer.value)
+        const submission = normalize(value)
 
-  switch (type) {
-    case InputExerciseType.InputNumberExactMatchChallenge:
-      return normalizeNumber(temp).replace(/\s/g, '')
-    case InputExerciseType.InputExpressionEqualMatchChallenge:
-      return A.parse(normalizeNumber(temp))
-    case InputExerciseType.InputStringNormalizedMatchChallenge:
-      return temp.toUpperCase()
-  }
-}
-
-function matchesInput(
-  field: { type: InputExerciseType; value: string },
-  input: string
-) {
-  try {
-    const solution = normalize(field.type, field.value)
-    const submission = normalize(field.type, input)
-
-    switch (field.type) {
-      case InputExerciseType.InputExpressionEqualMatchChallenge:
-        return (
-          (solution as A.Expression)
-            .subtract(submission as A.Expression)
-            .toString() === '0'
-        )
-      case InputExerciseType.InputNumberExactMatchChallenge:
-      case InputExerciseType.InputStringNormalizedMatchChallenge:
+        if (type === 'input-expression-equal-match-challenge' && solution) {
+          return (
+            (solution as A.Expression)
+              .subtract(submission as A.Expression)
+              .toString() === '0'
+          )
+        }
         return solution === submission
+      } catch (e) {
+        return false
+      }
+    })
+    const customFeedbackNode = filteredAnswers[0]?.feedback ?? null
+
+    const hasCorrectAnswer =
+      filteredAnswers.length > 0 && filteredAnswers[0].isCorrect
+
+    return {
+      correct: hasCorrectAnswer,
+      message: customFeedbackNode ?? (
+        <>{exStrings[hasCorrectAnswer ? 'correct' : 'wrong']}</>
+      ),
     }
-  } catch (err) {
-    // e.g. if user input could not be parsed
-    return false
   }
-}
 
-function normalizeNumber(numberText: string) {
-  return numberText.replace(/,/g, '.').replace(/^[+]/, '')
-}
+  function normalize(value: string) {
+    const _value = collapseWhitespace(value)
+    switch (type) {
+      case 'input-number-exact-match-challenge':
+        return normalizeNumber(_value).replace(/\s/g, '')
+      case 'input-expression-equal-match-challenge':
+        return A ? A.parse(normalizeNumber(_value)) : undefined
+      case 'input-string-normalized-match-challenge':
+        return _value.toUpperCase()
+    }
+  }
 
-function collapseWhitespace(text: string) {
-  return text.replace(/[\s\xa0]+/g, ' ').trim()
+  function collapseWhitespace(val: string): string {
+    return val.replace(/[\s\xa0]+/g, ' ').trim()
+  }
+
+  function normalizeNumber(val: string) {
+    return val.replace(/,/g, '.').replace(/^[+]/, '')
+  }
 }
